requirements.txt
cachetools==5.2.0
certifi==2022.6.15
chardet==5.0.0
click==8.1.3
Flask==2.2.2
google-auth==2.11.0
google-auth-oauthlib==0.5.2
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
oauthlib==3.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
urllib3==1.26.12
Werkzeug==2.2.2
pip==22.2.2